--- src/components/Loader/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Loader';
